#421 [refactor] Curious 서비스 분리
