+++ module-domain/src/main/java/com/mile/curious/service/CuriousRetriever.java
@@ -2,6 +2,10 @@ package com.mile.curious.service;
 
 import com.mile.curious.domain.Curious;
 import com.mile.curious.repository.CuriousRepository;
+import com.mile.exception.message.ErrorMessage;
+import com.mile.exception.model.ConflictException;
+import com.mile.exception.model.NotFoundException;
+import com.mile.post.domain.Post;
 import com.mile.writername.domain.WriterName;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
@@ -17,5 +21,21 @@ public class CuriousRetriever {
 
     public List<Curious> findAllByWriterName(final WriterName writerName) {
         return curiousRepository.findAllByWriterName(writerName);
+    }
+
+    public void checkCuriousExists(final Post post, final WriterName writerName) {
+        if (curiousRepository.existsByPostAndWriterName(post, writerName)) {
+            throw new ConflictException(ErrorMessage.CURIOUS_ALREADY_EXISTS_EXCEPTION);
+        }
+    }
+
+    public void checkCuriousNotExists(final Post post, final WriterName writerName) {
+        if (!curiousRepository.existsByPostAndWriterName(post, writerName)) {
+            throw new NotFoundException(ErrorMessage.CURIOUS_NOT_FOUND);
+        }
+    }
+
+    public boolean findCuriousExists(final Post post, final WriterName writerName) {
+        return curiousRepository.existsByPostAndWriterName(post, writerName);
     }
 }
