+++ module-domain/src/main/java/com/mile/curious/service/CuriousService.java
@@ -1,16 +1,10 @@
 package com.mile.curious.service;
 
-import com.mile.curious.domain.Curious;
-import com.mile.curious.repository.CuriousRepository;
 import com.mile.curious.service.dto.CuriousInfoResponse;
-import com.mile.exception.message.ErrorMessage;
-import com.mile.exception.model.ConflictException;
-import com.mile.exception.model.NotFoundException;
 import com.mile.post.domain.Post;
 import com.mile.writername.domain.WriterName;
 import com.mile.writername.service.WriterNameUpdator;
 import lombok.RequiredArgsConstructor;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,44 +13,29 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CuriousService {
 
-    private final CuriousRepository curiousRepository;
     private final WriterNameUpdator writerNameUpdator;
     private final CuriousRemover curiousRemover;
+    private final CuriousRetriever curiousRetriever;
+    private final CuriousCreator curiousCreator;
 
     public void deleteCurious(final Post post, final WriterName writerName) {
-        checkCuriousNotExists(post, writerName);
-        curiousRepository.delete(curiousRepository.findByPostAndWriterName(post, writerName));
+        curiousRetriever.checkCuriousNotExists(post, writerName);
+        curiousRemover.deleteCurious(post, writerName);
         post.decreaseCuriousCount();
         writerNameUpdator.decreaseTotalCuriousCountByWriterName(writerName);
     }
 
-    public void checkCuriousNotExists(final Post post, final WriterName writerName) {
-        if (!curiousRepository.existsByPostAndWriterName(post, writerName)) {
-            throw new NotFoundException(ErrorMessage.CURIOUS_NOT_FOUND);
-        }
-    }
-
     public void createCurious(final Post post, final WriterName writerName) {
-        checkCuriousExists(post, writerName);
-        try {
-            curiousRepository.save(Curious.create(post, writerName));
-        } catch (DataIntegrityViolationException e) {
-            throw new ConflictException(ErrorMessage.CURIOUS_ALREADY_EXISTS_EXCEPTION);
-        }
+        curiousRetriever.checkCuriousExists(post, writerName);
+        curiousCreator.createCurious(post, writerName);
         post.increaseCuriousCount();
         writerNameUpdator.increaseTotalCuriousCountByWriterName(writerName);
     }
 
-    public void checkCuriousExists(final Post post, final WriterName writerName) {
-        if (curiousRepository.existsByPostAndWriterName(post, writerName)) {
-            throw new ConflictException(ErrorMessage.CURIOUS_ALREADY_EXISTS_EXCEPTION);
-        }
-    }
 
     public CuriousInfoResponse getCuriousInfoOfPostAndWriterName(final Post post, final WriterName writerName) {
-        return CuriousInfoResponse.of(curiousRepository.existsByPostAndWriterName(post, writerName), post.getCuriousCount());
+        return CuriousInfoResponse.of(curiousRetriever.findCuriousExists(post, writerName), post.getCuriousCount());
     }
-
 
     public void deleteAllByPosts(
             final List<Post> posts
